requirements.txt
certifi==2026.2.25
cffi==1.15.1
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
cryptography==45.0.7
Flask==2.2.5
flask-sqlalchemy==3.0.5
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.21
python-dotenv==0.21.1
requests==2.31.0
sqlalchemy==2.0.48
typing-extensions==4.7.1
urllib3==2.0.7
waitress==2.1.2
Werkzeug==2.2.3
zipp==3.15.0
